requirements.txt
browser-use
langchain-google-genai
python-dotenv
fastapi
uvicorn
